skia: Fix a bug in clampx_nofilter_trans_S16_D32_DX
When scaling a line, in the case of the last pixel needing to be
duplicated, the beginning of the line is getting clobbered with wrong
value.  This is because dest buffer pointer is not moved forward.  Fix
that by adding pointer update line.

CRs-Fixed: 349571

(cherry picked from commit 3c6384cb47e6e09802944388335eea1b98c61e5c)

Change-Id: I3bc4756a7327579fb5271258612e44a8ae6eda25